Refactor by moving constants to another file
